requirements.txt
flask
openai
flask_caching
pymongo
pyautogen